FW: Fix NPE caused by changing Navigation Bar buttons
Changing NavBar buttons was causing an NPE, removal of the back button was causing
the System UI to FC on the i9100.

Method causing the problem is setNavigationIconHints, initial changed worked but would
cause the same error when opening the keyboard, or other dynamic NavBar targets
attempted to display.

This commit leaves the PAC method but also changes the inital method version and
helper to operate without the additional overriding parameter.

Change-Id: I4e89f12c2bd8361316571235887f4bd856680cb0