FW: Fix Notification Shortcuts
All credits to SlimRoms for their awesome work on the Notifications shortcuts, this commit corrects the PAC code to the SlimRoms 4.3 implementation.

Unable to cherry-pick a commit to correct our code issues but the original commit retains authorship and all credits for this code are solely attributed to the SlimRoms team, I have only corrected the PAC code port of this feature.

Change-Id: I418dba621cd98745cb6804a2bdd2ba447e580daf